Refactor AzureMonitorConvertor
