+++ sdk/monitor/Microsoft.OpenTelemetry.Exporter.AzureMonitor/tests/Benchmarks/TagObjectsBenchmarks.cs
@@ -12,7 +12,7 @@ namespace Microsoft.OpenTelemetry.Exporter.AzureMonitor.Benchmarks
     [MemoryDiagnoser]
     public class TagObjectsBenchmarks
     {
-        private AzureMonitorConverter.TagEnumerationState monitorTags;
+        private TagEnumerationState monitorTags;
         private IEnumerable<KeyValuePair<string, object>> tagObjects;
         private IEnumerable<KeyValuePair<string, object>> PartB_tagObjects;
         private IEnumerable<KeyValuePair<string, object>> PartC_tagObjects;
@@ -39,7 +39,7 @@ namespace Microsoft.OpenTelemetry.Exporter.AzureMonitor.Benchmarks
         [GlobalSetup]
         public void Setup()
         {
-            monitorTags = new AzureMonitorConverter.TagEnumerationState
+            monitorTags = new TagEnumerationState
             {
                 PartBTags = AzMonList.Initialize(),
                 PartCTags = AzMonList.Initialize()
